project folder renamed to momentum_calculator
