+++ momentum_calculator/calculator.py
